money: replace double with decimal for monetary values
Balance, Amount, BalanceAfterTransaction, and exchange-rate Value are now
decimal end to end (DataAccess models, Services models, UI models, the
IAccountRepository / IAccountService.UpdateBalance signatures, and the
service implementation). Repository readers drop the decimal.ToDouble
round-trip and read DECIMAL(18,2) directly. Writes for balance and
transaction amounts now use explicit SqlParameter with Precision=18,
Scale=2 instead of AddWithValue, which silently negotiates the wrong
precision for decimal values. Side fix: two latent ExecuteScalar calls
on UPDATE/INSERT statements (CreateAccount, UpdateBalance) are now
ExecuteNonQuery.

The DB schema is unchanged; this commit is purely about not laundering
DECIMAL through binary floating-point in the application tier.

+++ EBanking.DataAccess/Repositories/IAccountRepository.cs
@@ -7,7 +7,7 @@ namespace EBanking.DataAccess.Repositories
         void CreateAccount(Account model);
         List<Account> GetAccountsByUserId(int userId);
         Task<Account> GetAccountByAccountNumber(string accountNumber);
-        void UpdateBalance(double balance, string accountNumber);
+        void UpdateBalance(decimal balance, string accountNumber);
         bool IsValidAccount(string accountNumber);
     }
 }
