+++ EBanking.UI/Models/CurrencyExchangeModel.cs
@@ -5,8 +5,8 @@ namespace EBanking.UI.Models
 {
     public class CurrencyExchangeModel : BaseModel
     {
-        public double Amount { get; set; }
-        public double ConvertedValue { get; set; }
+        public decimal Amount { get; set; }
+        public decimal ConvertedValue { get; set; }
         public Services.Models.AccountModel Account { get; set; }
         public Services.Models.AccountModel SelectedAccount { get; set; }
         public List<Services.Models.AccountModel> UserAccounts { get; set; }
